src/lambda/requirements.txt
boto==2.43.0
beautifulsoup4==4.5.3
